implement --ignore-db-dirs option for xtrabackup.
This is meant to mimic the mysqld (5.6, mariadb) ignore-db-dirs
option where a comma separated list of dirs will be skipped.

This option has no default and ignores nothing out of the box. This
would usually be set to "lost+found" for cases where the MySQL datadir
is mounted on the root of a mountpoint on an ext* filesystem.

Fixes Launchpad bug #1272329